test: BlackjackGame 테스트 분리 + GameResult 테스트 추가
